ADD second iteration of the css practice html
revision include adding various content to the site. decided to try and play with placing an image and buttons. Not happy with the current button but acceptable practice. Unlikely to come back and edit again, as planned to press on to the following lesson. Expect no update from now on.